Завершает генератор и заканчивает модуляцию
