+++ js/generate-photo-gallery/data.js
@@ -1,0 +1,25 @@
+// Для генерации комментариев в галерее
+const commentMessages = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+// Для генерации имен коментариях
+const commentNames = ['Егор', 'Виктория', 'Марина', 'Алексей', 'Ева', 'Ксения', 'Дмитрий', 'Леонид', 'Георгий', 'Алисия', 'Никита', 'Игорь', 'Таисия', 'София', 'Сергей', 'Роман', 'Лев', 'Юлия', 'Артём'];
+
+// Для генерации описаний фотографий в Галерее
+const photoDescription = [
+  'Момент сохранен в этом фото',
+  'Фотография-это способ отразить мысли',
+  'В этом фото каждый может увидеть свой смысл',
+  'Фотографировать это искусство, но каждый ли может',
+  'Сохраняйте воспоминания, запечатляя их',
+  'Я очень много сил вложил в это фото',
+  'Я начинающий фотограф, оцените мое первое фото',
+];
+
+export {commentMessages, commentNames, photoDescription};
