+++ js/main.js
@@ -1,1 +1,1 @@
-import './generate-photo-gallery.mjs';
+import './generate-photo-gallery/generate-photo-gallery.js';
